requirements.txt
environs~=8.0.0
Flask>=1.0.2
Flask-Login>=0.4.1
Flask-RESTful>=0.3.8
Flask-SQLAlchemy>=2.4.1
requests>=2.25.1
SQLAlchemy>=1.3.23
SQLAlchemy-serializer>=1.3.4.4
WTForms>=2.2.1
Werkzeug>=1.0.1
yandex-money-sdk
Flask-WTF>=0.14.3
